[feat] #32 timeOptionIndex TimerTodo 변수로 추가

+++ app/src/main/java/org/android/bbangzip/presentation/ui/timer/navigation/TimerNavigation.kt
@@ -18,7 +18,7 @@ fun NavController.navigateToTimer(navOptions: NavOptions, shouldRestart: Boolean
 
 fun NavGraphBuilder.timerNavGraph(
     sharedViewModel: SharedViewModel,
-    navigateToTimerTodo: () -> Unit = {},
+    navigateToTimerTodo: (Int) -> Unit = {},
 ) {
     composable<BottomNavigationRoute.Timer> { backStackEntry ->
         val item = backStackEntry.toRoute<BottomNavigationRoute.Timer>()
